data_a - Age and Looks

The Age and Looks data set, data_a, contains data on age and appearance of fictional individuals. A possible question to investigate with this data set is whether there is a dependency between different variables of the data set.

data(data_a)

Format

A data frame with 40 rows and 6 variables:

bsex

biological sex

age

age of person

height

height of person

hair

hair colour

eyes

eye colour

glasses

info on whether person wears glasses

Source

Schmitz, D., & Esser, J. (2022). SfL: Statistics for Linguistics. R package version 0.3. URL: https://github.com/dosc91/SfL