data_c - Incomplete Neutralization

The Incomplete Neutralization data set, data_c, contains data on the phonological phenomenon of incomplete neutralization. The values are fictional, but closely related to real data, i.e. the data by Roettger et al. (2014). A possible question to investigate with this data set is whether vowels are shorter when preceding voiceless word-final consonants.

data_c

Format

A data frame with 300 rows and 9 variables:

subject

biological sex

gender

gender of subject

item

word produced

voicing

voicing of final consonant

vowel

vowel preceding the final consonant

place

place of articulation of final consonants

accent_type

nuclear or prenuclear accent on word

prosodic_boundary

info on whether word is followed by a boundary

vowel_dur

duration of vowel preceding the final consonant

Source

Roettger, T. B., Winter, B., Kirby, J., Grawunder, S., & Grice, M. (2014). Assessing incomplete neutralization of final devoicing in German. Journal of Phonetics, 43: 11- 25.