data_t - Tonal Alignment in German

The Tonal Alignment in German data set, data_t, contains data on tonal alignment in German. The values are fictional, but closely related to real data, i.e. the data by Schmitz (2018). A possible question to investigate with this data set is whether the frequency of low points differs by syllable structure.

data_t

Format

A data frame with 544 rows and 15 variables:

speaker

speaker ID

word_dur

word duration

vowel

vowel of word

sylnum

number of syllables of word

sylstr

syllable structure of first syllable of word

C1_dur

duration of consonant preceding the vowel

V1_dur

vowel duration of first syllable vowel

C2_dur

duration of first consonant following the vowel within the same syllable

C3_dur

duration of second consonant following the vowel within the same syllable

W_dur

duration of onset consonant of 2nd syllable

A1_dur

duration of vowel of 2nd syllable

L_dur

duration of onset consonant of 3rd syllable

A2_dur

duration of vowel of 3rd syllable

L_time

point in time when lowest F0 value, i.e. L, occured

L_freq

frequency of lowest F0 value, i.e. L

Source

This data set is a shortened version of: Schmitz, D. (2018). Tonal Alignment and Segmental Composition in German. Master's Thesis. Universität zu Köln, Köln, Germany.