data_v - Vowel Shortening in German

The Vowel Shortening in German data set, data_v, contains data on the duration of vowels in open, simplex, and complex syllables in German. The data is fictional, but closely related to the original data used by Schmitz et al. (2018). A possible question to investigate with this data set is whether vowel duration differs by syllable structure.

data_v

Format

A data frame with 448 rows and 6 variables:

speaker

speaker ID

word

pseudoword produced, contains the vowel

duration

vowel duration

structure

coda structure

vowel

one of five vowels

rate

speech rate

Source

Schmitz, D., Hae-Eun C., & Niemann, H. (2018). Vowel shortening in German as a function of syllable structure. In Proceedings 13. Phonetik und Phonologie Tagung (P&P13), 181–184. Berlin.